+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Vuex implementation fixes
 - Increase/decrease combatant's advantage
 - Automatically set advantage of character to 0 if it takes any damage
+- Add filter for unique/non-unique NPCs in bestiary
+- Add state versioning and backwards compatibility
 
 ##[0.4.0]
 - Further implement Vuex, major refactor of use of variables and saving of state